# packages/shared-contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: EduNexus AI API Contract
  version: 1.2.0
  description: |
    EduNexus AI MVP 接口契约（v1.2.0）。
    本文件是前端、后端、AI 服务联调与代码生成的唯一接口基线。
    所有响应均遵循 ApiResponse 包装，data 字段类型见各接口描述与 schemas 定义。
    枚举值权威定义见 04-领域模型 §5，本文件仅做引用。
    API -> AI Service 的内部契约不在本文件定义，见 12-内部服务契约与异步任务规范。
    
    ## 接口 → data 类型速查
    | 接口 | data 类型 |
    |------|----------|
    | POST /auth/register | UserVO |
    | POST /auth/login | LoginData |
    | POST /auth/logout | null |
    | POST /auth/refresh | RefreshData |
    | GET /auth/me | UserVO |
    | POST /student/chat/session | ChatSessionVO |
    | GET /student/chat/sessions | PagedData<ChatSessionVO> |
    | GET /student/chat/session/{id} | ChatSessionDetailVO |
    | DELETE /student/chat/session/{id} | null |
    | POST /student/chat/session/{id}/message | ChatReplyVO |
    | GET /student/exercise/questions | PagedData<QuestionVO> |
    | POST /student/exercise/submit | ExerciseResultVO |
    | GET /student/exercise/{id}/analysis | ExerciseAnalysisVO |
    | GET /student/exercise/wrong-questions | PagedData<WrongBookEntryVO> |
    | DELETE /student/exercise/wrong-questions/{id} | null |
    | GET /student/exercise/records | PagedData<ExerciseRecordVO> |
    | POST /student/ai-questions/generate | AiQuestionGenerateResultVO |
    | GET /student/ai-questions | PagedData<AiQuestionSessionVO> |
    | POST /student/ai-questions/submit | AiQuestionResultVO |
...
    | GET /student/profile/weak-points | WeakPointVO[] |
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /student/chat/sessions:
    get:
      tags: [Student-Chat]
      summary: 获取聊天会话列表
      operationId: listChatSessions
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: 获取成功。data → PagedData，content 元素为 ChatSessionVO
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /student/chat/session/{sessionId}:
    get:
      tags: [Student-Chat]
      summary: 获取会话详情
      operationId: getChatSessionDetail
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: 获取成功。data → ChatSessionDetailVO（含 messages 列表）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Student-Chat]
      summary: 删除会话（软删除）
      operationId: deleteChatSession
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: 删除成功。data → null
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
  /student/chat/session/{sessionId}/message:
    post:
      tags: [Student-Chat]
      summary: 向会话发送消息（支持 SSE）
      operationId: sendChatMessage
      parameters:
        - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatSendRequest'
      responses:
        '200':
          description: 响应成功。data → ChatReplyVO（JSON）或 SSE 流（text/event-stream）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            text/event-stream:
              schema:
                type: string
                example: |
                  data: {"delta":"力的公式是 F=ma","citations":[{"title":"physics_ch3.pdf","score":0.92}]}
                  data: [DONE]
        '403':
          $ref: '#/components/responses/Forbidden'
  /student/exercise/questions:
    get:
      tags: [Student-Exercise]
      summary: 获取练习题列表
      operationId: listExerciseQuestions
      parameters:
        - name: subject
          in: query
          schema:
            type: string
        - name: difficulty
          in: query
          schema:
            type: string
            enum: [EASY, MEDIUM, HARD]
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: 获取成功。data → PagedData，content 元素为 QuestionVO
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /student/exercise/submit:
    post:
      tags: [Student-Exercise]
      summary: 提交练习答案并自动判分
      operationId: submitExerciseAnswers
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExerciseSubmitRequest'
      responses:
        '200':
          description: 提交成功。data → ExerciseResultVO（含逐题判分结果）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /student/exercise/{recordId}/analysis:
    get:
      tags: [Student-Exercise]
      summary: 查看练习解析
      operationId: getExerciseAnalysis
      parameters:
        - $ref: '#/components/parameters/RecordId'
      responses:
        '200':
          description: 查询成功。data → ExerciseAnalysisVO（含逐题解析与教师建议）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /student/exercise/wrong-questions:
    get:
      tags: [Student-Exercise]
      summary: 获取错题本
      operationId: listWrongQuestions
      parameters:
        - name: subject
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVE, MASTERED]
            default: ACTIVE
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: 查询成功。data → PagedData，content 元素为 WrongBookEntryVO
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /student/exercise/wrong-questions/{questionId}:
    delete:
      tags: [Student-Exercise]
      summary: 从错题本移除（标记已掌握）
      operationId: removeWrongQuestion
      parameters:
        - $ref: '#/components/parameters/QuestionId'
      responses:
        '200':
          description: 移除成功。data → null
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /student/exercise/records:
    get:
      tags: [Student-Exercise]
      summary: 获取做题记录
      operationId: listExerciseRecords
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: 查询成功。data → PagedData，content 元素为 ExerciseRecordVO
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /student/ai-questions/generate:
    post:
      tags: [Student-AIQuestions]
      summary: 生成 AI 个性化题目
      operationId: generateAiQuestions
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AiQuestionGenerateRequest'
      responses:
        '200':
          description: 生成成功。data → AiQuestionGenerateResultVO（含生成题目列表）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /student/ai-questions:
    get:
      tags: [Student-AIQuestions]
      summary: 获取 AI 出题历史
      operationId: listAiQuestionSessions
      parameters:
        - name: subject
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: 查询成功。data → PagedData，content 元素为 AiQuestionSessionVO
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /student/ai-questions/submit:
    post:
      tags: [Student-AIQuestions]
      summary: 提交 AI 题目答案
      operationId: submitAiQuestions
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AiQuestionSubmitRequest'
      responses:
        '200':
          description: 提交成功。data → AiQuestionResultVO
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /student/ai-questions/{recordId}/analysis:
    get:
      tags: [Student-AIQuestions]
      summary: 获取 AI 题目解析
      operationId: getAiQuestionAnalysis
      parameters:
        - $ref: '#/components/parameters/RecordId'
      responses:
        '200':
          description: 查询成功。data → AiQuestionAnalysisVO
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    Size:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    RecordId:
      name: recordId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    QuestionId:
      name: questionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: 幂等键（建议 UUID），用于防止重复提交导致的重复写入
      schema:
        type: string
        minLength: 8
        maxLength: 128
  responses:
...
    Forbidden:
      description: 无权限
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            code: 403
            message: 无权限访问
            data: null
            traceId: trace-xxx
            timestamp: '2026-02-26T10:00:00Z'
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            code: 404
            message: 资源不存在
            data: null
            traceId: trace-xxx
            timestamp: '2026-02-26T10:00:00Z'
  schemas:
...
    ApiResponse:
      type: object
      required: [code, message, timestamp]
      description: 统一响应包装。具体接口通过 allOf 指定 data 类型。
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          description: 具体业务数据（由各接口 Schema 定义）
        traceId:
          type: string
        timestamp:
          type: string
          format: date-time
        errorCode:
          type: string
          nullable: true
          description: 结构化错误码（失败响应可用）
...
    ChatSendRequest:
      type: object
      required: [message]
      properties:
        message:
          type: string
          minLength: 1
          maxLength: 4000
    ExerciseSubmitRequest:
      type: object
      required: [answers]
      properties:
        answers:
          type: array
          minItems: 1
          items:
            type: object
            required: [questionId, userAnswer]
            properties:
              questionId:
                type: string
                format: uuid
              userAnswer:
                type: string
    AiQuestionGenerateRequest:
      type: object
      required: [count, subject]
      properties:
        count:
          type: integer
          minimum: 1
          maximum: 20
        subject:
          type: string
        difficulty:
          type: string
          description: 枚举定义见 04-领域模型 §5
          enum: [EASY, MEDIUM, HARD]
        conceptTags:
          type: array
          items:
            type: string
    AiQuestionSubmitRequest:
      type: object
      required: [sessionId, answers]
      properties:
        sessionId:
          type: string
          format: uuid
        answers:
          type: array
          minItems: 1
          items:
            type: object
            required: [questionId, userAnswer]
            properties:
              questionId:
                type: string
                format: uuid
              userAnswer:
                type: string
...